+++ core/php/AbeilleParser-Xiaomi.php
@@ -25,7 +25,7 @@
     // Based on https://github.com/dresden-elektronik/deconz-rest-plugin/wiki/Xiaomi-manufacturer-specific-clusters%2C-attributes-and-attribute-reporting
     // Could be cluster 0000, attr FF01, type 42 (character string)
     // Could be cluster FCC0, attr 00F7, type 41 (octet string)
-    function xiaomiDecodeTags($net, $addr, $pl, &$attrReportN = null) {
+    function xiaomiDecodeTags($net, $addr, $pl, &$attrReportN = null, &$toMon = null) {
         // $tagsList = array(
         //     "01-21" => array( "desc" => "Battery-Volt" ),
         //     "03-28" => array( "desc" => "Device-Temp" ),
@@ -121,11 +121,15 @@
                     $attrReportN[] = array( "name" => $map['info'], "value" => $value2 );
             } else
                 $mapTxt = ' (ignored)';
-            parserLog('debug', '  Tag='.$tagId.', Type='.$typeId.'/'.$type['short'].', ValueHex='.$valueHex.' => '.$value.$mapTxt);
+            $m = '  Tag='.$tagId.', Type='.$typeId.'/'.$type['short'].', ValueHex='.$valueHex.' => '.$value.$mapTxt;
+            parserLog('debug', $m);
+            if ($toMon !== null)
+                $toMon[] = $m;
+
         }
     }
 
-    function xiaomiReportAttributes($net, $addr, $pl, &$attrReportN = null) {
+    function xiaomiReportAttributes($net, $addr, $pl, &$attrReportN = null, &$toMon = null) {
         $l = strlen($pl);
         for ( ; $l > 0; ) {
             $attrId = substr($pl, 2, 2).substr($pl, 0, 2);
@@ -157,7 +161,7 @@
 
                 parserLog('debug', '  Xiaomi proprietary (Door Sensor)');
                 $attributesReportN = [];
-                xiaomiDecodeTags($net, $addr, $pl2, $attrReportN);
+                xiaomiDecodeTags($net, $addr, $pl2, $attrReportN, $toMon);
 
                 // Previous code. For info only
                 // // $voltage        = hexdec(substr($payload, 24 + 2 * 2 + 2, 2).substr($payload, 24 + 2 * 2, 2));
@@ -179,7 +183,7 @@
                 parserLog('debug', '  Xiaomi proprietary (Capteur Presence V2)');
 
                 // For info until activation
-                xiaomiDecodeTags($net, $addr, $pl2);
+                xiaomiDecodeTags($net, $addr, $pl2, null, $toMon);
                 // For info until activation
 
                 // $voltage = hexdec(substr($pl2, 28+2, 2).substr($pl2, 28, 2));
@@ -199,7 +203,7 @@
                 parserLog('debug', '  Xiaomi proprietary (Water leak sensor)');
 
                 // For info until activation
-                xiaomiDecodeTags($net, $addr, $pl2);
+                xiaomiDecodeTags($net, $addr, $pl2, null, $toMon);
                 // For info until activation
 
                 // $voltage = hexdec(substr($pl2, 24 + 2 * 2 + 2, 2).substr($pl2, 24 + 2 * 2, 2));
@@ -217,7 +221,7 @@
 
                 parserLog('debug', '  Xiaomi proprietary (Temp square sensor)');
                 $attributesReportN = [];
-                xiaomiDecodeTags($net, $addr, $pl2, $attrReportN);
+                xiaomiDecodeTags($net, $addr, $pl2, $attrReportN, $toMon);
 
                 // Previous code. For info only
                 $voltage        = hexdec(substr($pl2, 2 * 2 + 2, 2).substr($pl2, 2 * 2, 2));
@@ -238,7 +242,7 @@
             //
 
             else if (($attrId == "00F7") && ($attrType == "41")) {
-                xiaomiDecodeTags($net, $addr, $pl2, $attrReportN);
+                xiaomiDecodeTags($net, $addr, $pl2, $attrReportN, $toMon);
             }
 
             // break; // Currently supporting only 1 attribut.