+++ docs/fr_FR/Changelog.rst
@@ -2,6 +2,7 @@ ChangeLog
 =========
 
 - Legrand micromodule switch: Mise-à-jour modele et image.
+- Interne: Parser: Amélioration mess monitor cas Xiaomi.
 
 230106-BETA-2
 -------------